chore: move fixture script to fixtures
