+++ packages/launchql-client/src/meta-object/convert.js
@@ -1,6 +1,3 @@
-import path from 'path';
-import fs from 'fs';
-
 export function convertFromMetaSchema(metaSchema) {
   const {
     _meta: { tables }
@@ -66,23 +63,4 @@ function pickField(field) {
     name: field.name,
     type: field.type
   };
-}
-
-function generateMetaObjectFixture() {
-  const inDir = path.resolve(
-    __dirname,
-    '../../__fixtures__/api/meta-schema.json'
-  );
-  const outDir = path.resolve(
-    __dirname,
-    '../../__fixtures__/api/meta-obj.json'
-  );
-  fs.readFile(inDir, { encoding: 'utf8' }, (err, data) => {
-    if (err) return console.log(err);
-    const converted = convertFromMetaSchema(JSON.parse(data));
-    fs.writeFile(outDir, JSON.stringify(converted), (err) => {
-      if (err) return console.log(err);
-      console.log('DONE');
-    });
-  });
 }
